Re-organize folder structure and hierarchy for DS2.
