+++ deep_speech_2/cloud/pcloud_train.sh
@@ -1,3 +1,5 @@
+#! /usr/bin/bash
+
 TRAIN_MANIFEST=$1
 DEV_MANIFEST=$2
 MODEL_PATH=$3
@@ -14,11 +16,29 @@ python ./cloud/split_data.py \
 --out_manifest_path='/local.manifest.dev'
 
 python -u train.py \
---batch_size=$BATCH_SIZE \
---use_gpu=1 \
+--batch_size=${BATCH_SIZE} \
 --trainer_count=${NUM_GPU} \
---num_threads_data=${NUM_GPU} \
+--num_passes=200 \
+--num_proc_data=${NUM_GPU} \
+--num_conv_layers=2 \
+--num_rnn_layers=3 \
+--rnn_layer_size=2048 \
+--num_iter_print=100 \
+--learning_rate=5e-4 \
+--max_duration=27.0 \
+--min_duration=0.0 \
+--use_sortagrad=True \
+--use_gru=False \
+--use_gpu=True \
 --is_local=${IS_LOCAL} \
---train_manifest_path='/local.manifest.train' \
---dev_manifest_path='/local.manifest.dev' \
---output_model_dir=${MODEL_PATH} 2>&1 | tee ./log/train.log
+--share_rnn_weights=True \
+--train_manifest='/local.manifest.train' \
+--dev_manifest='/local.manifest.dev' \
+--mean_std_path='data/librispeech/mean_std.npz' \
+--vocab_path='data/librispeech/eng_vocab.txt' \
+--output_model_dir='./checkpoints' \
+--output_model_dir=${MODEL_PATH} \
+--augment_conf_path='conf/augmentation.config' \
+--specgram_type='linear' \
+--shuffle_method='batch_shuffle_clipped' \
+2>&1 | tee ./log/train.log
