+++ deep_speech_2/utils/utility.py
